Reorganize project structure and update imports
