+++ backend/__init__.py
@@ -1,0 +1,1 @@
+"""Paquete que contiene el backend FastAPI."""
