+++ backend/models/task.py
@@ -1,6 +1,7 @@
+"""Definiciones de modelo ORM para tareas."""
 from sqlalchemy import Column, Integer, String
 
-from ..database import Base
+from ..config.database import Base
 
 class Task(Base):
     """Simple task model used for demo purposes."""
