+++ tests/test_tasks.py
@@ -3,8 +3,8 @@
 import os
 from fastapi.testclient import TestClient
 
-from app.main import app
-from app.database import Base, engine
+from backend.main import app
+from backend.config.database import Base, engine
 
 client = TestClient(app)
